novo commit
ultimo código com busca CEP

+++ src/extintor/extintor.controller.ts
@@ -3,7 +3,6 @@ import {
   Controller,
   Get,
   Post,
-  Put,
   Body,
   Patch,
   Param,
@@ -15,7 +14,6 @@ import { ExtintorService } from './extintor.service';
 import { CreateExtintorDto } from './dto/create-extintor.dto';
 import { UpdateExtintorDto } from './dto/update-extintor.dto';
 import { AuthService } from 'src/auth/auth.service';
-import { query } from 'express';
 
 @Controller('extintor')
 export class ExtintorController {
@@ -25,7 +23,12 @@ export class ExtintorController {
   ) {}
 
   @Post()
-  create(@Body() createExtintorDto: CreateExtintorDto) {
+  create(
+    @Headers('x-api-token') token: string,
+    @Body() createExtintorDto: CreateExtintorDto,
+  ) {
+    if (!token) throw new UnauthorizedException('Token não enviado!');
+    this.authService.validadeToken(token);
     return this.extintorService.create(createExtintorDto);
   }
 
@@ -37,36 +40,49 @@ export class ExtintorController {
     @Query('preco') preco?: string,
     @Query('validade') validade?: string,
     @Query('peso') peso?: string,
+    @Query('id') id?: string,
   ) {
     if (!token) throw new UnauthorizedException('Token não enviado!');
-
     this.authService.validadeToken(token);
 
     return this.extintorService.findAll(
       nome,
       classe,
-      preco ? parseInt(preco): undefined,
-      peso ? parseInt(peso): undefined,
+      preco ? parseFloat(preco) : undefined,
+      peso ? parseFloat(peso) : undefined,
       validade,
-      
+    id ? parseInt(id) : undefined,
     );
   }
 
   @Get(':id')
-  findOne(@Param('id') id: string) {
+  findOne(
+    @Param('id') id: string,
+    @Headers('x-api-token') token: string,
+  ) {
+    if (!token) throw new UnauthorizedException('Token não enviado!');
+    this.authService.validadeToken(token);
     return this.extintorService.findOne(+id);
   }
 
   @Patch(':id')
   update(
+    @Headers('x-api-token') token: string,
     @Param('id') id: string,
     @Body() updateExtintorDto: UpdateExtintorDto,
   ) {
+    if (!token) throw new UnauthorizedException('Token não enviado!');
+    this.authService.validadeToken(token);
     return this.extintorService.update(+id, updateExtintorDto);
   }
 
   @Delete(':id')
-  remove(@Param('id') id: string) {
+  remove(
+    @Headers('x-api-token') token: string,
+    @Param('id') id: string,
+  ) {
+    if (!token) throw new UnauthorizedException('Token não enviado!');
+    this.authService.validadeToken(token);
     return this.extintorService.remove(+id);
   }
 }
